Refactor: 에러 처리방식 통일 및 application.yml환경변수설정(#21)
