+++ src/main/java/com/partnerd/apiPaylaod/exception/handler/CategoryHandler.java
@@ -1,0 +1,9 @@
+package com.partnerd.apiPaylaod.exception.handler;
+
+import com.partnerd.apiPaylaod.code.BaseErrorCode;
+import com.partnerd.apiPaylaod.exception.GeneralException;
+
+public class CategoryHandler extends GeneralException {
+
+    public CategoryHandler(BaseErrorCode errorCode){super (errorCode); }
+}
